pyproject.toml
[tool.poetry]
name = "password_comp"
version = "0.1.0"
description = ""
authors = ["Lisstrange <nikita.a.lisovik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.1"
optuna = "^2.9.1"
numpy = "^1.21.1"
Flask = "^2.0.1"
scikit-learn = "^0.24.2"
jupyter = "^1.0.0"
notebook = "^6.4.1"
nltk = "^3.6.2"
lightgbm = "^3.2.1"
pathlib = "^1.0.1"
python-box = "^5.4.1"
Flask-Bootstrap = "^3.3.7"

[tool.poetry.dev-dependencies]
seaborn = "^0.11.1"
matplotlib = "^3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
